+++ CHANGES.md
@@ -1,2 +1,3 @@
 - Update for patch 10.1.5
 - Update instance and trinket lists
+- Fix loot alert using corrupted template